Player Crouching MK.2 Pt 3
Finished toggleable player crouching. Does not uncrouch if the standing height is shorter than the ceiling. Model and collider works perfectly.
