--- Assets/Scripts/Player/PlayerCrouch.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9e178633105855d409d68c29cd6c4ba0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
